Fix ellipsis and bump to 0.8.8
Previously, text entered as "well.. this is awkward\nheyyy" would be weird because it would duplicate paragraphs. To explain is complicated. When RedCloth ran the text, it would attempt to see if there was an html function named the tag of the extended paragraph, in the previous it would look for :well, if not it would run a fallback. The fallback text itself was correctly stored in a 'reg' named 'fallback' which contained the text of the code ('well.. '). The problem here was that when it ran the fallback code, it would use the fallback text itself as temporary storage for the entire text. Except that it was modifying the permanent fallback text in a call-by-value 'regs' which only lived in the function. Thus when the second paragraph rolled around, the fallback not only wasn't cleared, it contained the text of all that happened previous, thus duplicating paragraphs. This can be resolved by force clearing out "fallback" text, thus it won't be used in subsequent paragraphs.

tl;dr: pointer math is hard.